requirements.txt
# GUI Framework
PyQt5>=5.15.0

# PDF Processing
PyMuPDF>=1.23.0

# DOCX Processing
python-docx>=1.1.0

# Property-based Testing
hypothesis>=6.0.0

# Unit Testing
pytest>=7.0.0
pytest-qt>=4.0.0

# Packaging
pyinstaller>=6.0.0
